updated for version 7.4.484 Problem:    Compiler warning on MS-Windows. (Ken Takata) Solution:   Add type cast.

+++ src/getchar.c
@@ -4530,7 +4530,7 @@ check_abbr(c, ptr, col, mincol)
 	    if (q != NULL)
 	    {
 		vim_unescape_csi(q);
-		qlen = STRLEN(q);
+		qlen = (int)STRLEN(q);
 		vim_free(q);
 	    }
 	}
